requirements.txt
asyncpg==0.30.0
backports.tarfile==1.2.0
greenlet==3.2.4
grpcio-status==1.71.2
importlib-metadata==8.0.0
pgvector==0.4.1
jaraco.collections==5.1.0
langchain-community==0.3.29
langchain-core>=0.3.0
langchain-groq==0.3.7
langdetect==1.0.9
langgraph-api==0.4.20
langgraph-cli==0.4.0
langsmith>=0.1.0
pip-chill==1.0.3
platformdirs==4.2.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pytest==8.4.2
pytest-asyncio==0.24.0
python-multipart==0.0.20
rapidfuzz==3.14.1
sentence-transformers==5.1.0
tomli==2.0.1
SQLAlchemy==2.0.43
torch==2.8.0
python-dotenv==1.1.1